myAwesomeCart/mac/requirements.txt
django-ckeditor 6.7.1